# server/openapi.json
{
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationErrorItem"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "detail"
                ],
                "type": "object"
            },
            "HttpErrorOut": {
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "detail": {
                        "type": "string"
                    }
                },
                "required": [
                    "code",
                    "detail"
                ],
                "title": "HttpError",
                "type": "object"
            },
...
            "MessageOut": {
                "properties": {
                    "channels": {
                        "description": "List of free-form identifiers that endpoints can filter by",
                        "example": [
                            "project_123",
                            "group_2"
                        ],
                        "items": {
                            "example": "project_1337",
                            "maxLength": 128,
                            "pattern": "^[a-zA-Z0-9\\-_.]+$",
                            "type": "string"
                        },
                        "maxItems": 5,
                        "minItems": 1,
                        "nullable": true,
                        "type": "array",
                        "uniqueItems": true
                    },
                    "eventId": {
                        "description": "Optional unique identifier for the message",
                        "example": "unique-msg-identifier",
                        "maxLength": 256,
                        "minLength": 1,
                        "nullable": true,
                        "pattern": "^[a-zA-Z0-9\\-_.]+$",
                        "type": "string"
                    },
                    "eventType": {
                        "example": "user.signup",
                        "maxLength": 256,
                        "pattern": "^[a-zA-Z0-9\\-_.]+$",
                        "type": "string"
                    },
                    "id": {
                        "example": "msg_1srOrx2ZWZBpBUvZwXKQmoEYga2",
                        "type": "string"
                    },
                    "payload": {
                        "example": {
                            "email": "test@example.com",
                            "username": "test_user"
                        },
                        "type": "object"
                    },
                    "timestamp": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "eventType",
                    "id",
                    "payload",
                    "timestamp"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "description": "HTTP Bearer token passed in the `Authorization` header",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "",
        "title": "Svix API",
        "version": "1.1.1",
        "x-logo": {
            "altText": "Svix Logo",
            "url": "https://www.svix.com/static/img/brand-padded.svg"
        }
    },
    "openapi": "3.0.2",
    "paths": {
...
        "/api/v1/app/{app_id}/msg/{msg_id}": {
            "get": {
                "description": "Get a message by its ID or eventID.",
                "operationId": "v1.message.get",
                "parameters": [
                    {
                        "in": "path",
                        "name": "app_id",
                        "required": true,
                        "schema": {
                            "example": "unique-app-identifier",
                            "maxLength": 256,
                            "minLength": 1,
                            "pattern": "^[a-zA-Z0-9\\-_.]+$",
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "path",
                        "name": "msg_id",
                        "required": true,
                        "schema": {
                            "example": "unique-msg-identifier",
                            "maxLength": 256,
                            "minLength": 1,
                            "pattern": "^[a-zA-Z0-9\\-_.]+$",
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "description": "When `true` message payloads are included in the response",
                        "in": "query",
                        "name": "with_content",
                        "schema": {
                            "default": true,
                            "description": "When `true` message payloads are included in the response",
                            "type": "boolean"
                        },
                        "style": "form"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageOut"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorOut"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorOut"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorOut"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorOut"
                                }
                            }
                        },
                        "description": "Conflict"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorOut"
                                }
                            }
                        },
                        "description": "Too Many Requests"
                    }
                },
                "summary": "Get Message",
                "tags": [
                    "Message"
                ]
            }
        },
...
        "/api/v1/app/{app_id}/msg/{msg_id}/attempt/{attempt_id}/content": {
            "delete": {
                "description": "Deletes the given attempt's response body. Useful when an endpoint accidentally returned sensitive content.",
                "operationId": "v1.message-attempt.expunge-content",
                "parameters": [
                    {
                        "in": "path",
                        "name": "app_id",
                        "required": true,
                        "schema": {
                            "example": "unique-app-identifier",
                            "maxLength": 256,
                            "minLength": 1,
                            "pattern": "^[a-zA-Z0-9\\-_.]+$",
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "path",
                        "name": "msg_id",
                        "required": true,
                        "schema": {
                            "example": "unique-msg-identifier",
                            "maxLength": 256,
                            "minLength": 1,
                            "pattern": "^[a-zA-Z0-9\\-_.]+$",
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "path",
                        "name": "attempt_id",
                        "required": true,
                        "schema": {
                            "example": "atmpt_1srOrx2ZWZBpBUvZwXKQmoEYga2",
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorOut"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorOut"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorOut"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorOut"
                                }
                            }
                        },
                        "description": "Conflict"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorOut"
                                }
                            }
                        },
                        "description": "Too Many Requests"
                    }
                },
                "summary": "Expunge Attempt Content",
                "tags": [
                    "Message Attempt"
                ]
            }
        },
...
    },
...
}